refactor(inspect): split inspector_overlay.cpp by concern (P10-2)
inspector_overlay.cpp had grown to 2,834 lines as the inspector
gained Phases 2-6. Roadmap P10-2 splits the per-feature overlay
clusters into sibling TUs, following the motion_geometry.cpp
extraction pattern: sibling .cpp files plus a PRIVATE internal
header for the file-local state that crosses the new TU boundary.

Three cohesive, self-contained feature clusters move out:
  inspector_overlay_field_edit.cpp  (302) — Phase 3b live-editable
    box-model fields: editable_field_at, read/write_field_value,
    begin/commit/cancel_field_edit, apply_edit_buffer_to_view,
    handle_edit_key + the file-local key_to_char helper.
  inspector_overlay_zoom.cpp        (284) — Phase 3e 20x zoom loupe:
    set_zoom_active/factor, resolve_view_color_at, update_zoom_sample,
    paint_zoom_panel.
  inspector_overlay_pass_viewer.cpp (242) — Phase 6.1 per-pass render
    attribution: capture_pass_frame, pass_attribution,
    paint_pass_attribution + the file-local kPassTypeNames/Colors
    tables.

The overlay's shared color palette (29 former file-scope statics)
relocates to a new private inspector_overlay_internal.hpp (under
inspect/src/, NOT the public inspect/include/ tree) as inline
constants so every overlay TU sees one definition. Structural
layout constants were already static-constexpr members of
InspectorOverlay, so the split TUs reach them through the public
header unchanged. Function-local color constants stay with their
(unmoved) clusters.

Pure mechanical move — extracted code is byte-identical, zero
behavior change. The eyedropper / drift / reconcile / tweaks /
atlas clusters stay in inspector_overlay.cpp because they share
helpers (collect_anchor_ids, abbreviate_anchor, color_to_hex)
across cluster boundaries; extracting them cleanly is out of scope
for this strictly-mechanical slice.

inspector_overlay.cpp: 2,834 -> 2,071 (-763).

Verified: pulp-inspect builds + links clean (with Skia);
pulp-test-inspector passes (816 assertions / 150 test cases).
