+++ inspect/CMakeLists.txt
@@ -9,6 +9,9 @@ target_include_directories(pulp-inspect
 
 target_sources(pulp-inspect PRIVATE
     src/inspector_overlay.cpp
+    src/inspector_overlay_field_edit.cpp
+    src/inspector_overlay_zoom.cpp
+    src/inspector_overlay_pass_viewer.cpp
     src/inspector_window.cpp
     src/state_inspector.cpp
     src/console_capture.cpp
