+++ inspect/src/inspector_overlay.cpp
@@ -1,4 +1,13 @@
 // inspector_overlay.cpp — Visual inspector overlay implementation
+//
+// Per-feature overlay clusters live in sibling TUs (roadmap P10-2):
+//   inspector_overlay_field_edit.cpp   — Phase 3b live-editable fields
+//   inspector_overlay_zoom.cpp         — Phase 3e 20× zoom loupe
+//   inspector_overlay_pass_viewer.cpp  — Phase 6.1 pass-attribution viewer
+// Shared overlay color constants are declared in
+// inspector_overlay_internal.hpp.
+
+#include "inspector_overlay_internal.hpp"
 
 #include <pulp/inspect/inspector_overlay.hpp>
 #include <pulp/inspect/tweak_store.hpp>
@@ -15,45 +24,6 @@
 #include <cmath>
 
 namespace pulp::inspect {
-
-// ── Colors ──────────────────────────────────────────────────────────────────
-
-static const Color kHighlightFill    = Color::rgba(0.25f, 0.5f, 1.0f, 0.12f);
-static const Color kHighlightStroke  = Color::rgba(0.25f, 0.5f, 1.0f, 0.7f);
-static const Color kSelectedFill     = Color::rgba(1.0f, 0.5f, 0.0f, 0.12f);
-static const Color kSelectedStroke   = Color::rgba(1.0f, 0.5f, 0.0f, 0.7f);
-static const Color kPanelBg          = Color::rgba(0.08f, 0.08f, 0.1f, 0.94f);
-static const Color kPanelText        = Color::rgba(0.85f, 0.85f, 0.9f, 1.0f);
-static const Color kPanelDim         = Color::rgba(0.5f, 0.5f, 0.55f, 1.0f);
-static const Color kPanelHighlight   = Color::rgba(0.25f, 0.5f, 1.0f, 0.25f);
-static const Color kTreeSelected     = Color::rgba(1.0f, 0.5f, 0.0f, 0.3f);
-static const Color kDistanceLine     = Color::rgba(1.0f, 0.2f, 0.3f, 0.8f);
-static const Color kPaddingColor     = Color::rgba(0.2f, 0.8f, 0.3f, 0.15f);
-static const Color kMarginColor      = Color::rgba(1.0f, 0.6f, 0.1f, 0.15f);
-static const Color kStatsBg          = Color::rgba(0.0f, 0.0f, 0.0f, 0.7f);
-static const Color kStatsText        = Color::rgba(0.6f, 1.0f, 0.6f, 1.0f);
-static const Color kStatsWarn        = Color::rgba(1.0f, 0.4f, 0.3f, 1.0f);
-
-// Phase 3b — editable-field visual treatment
-static const Color kFieldEditCaret   = Color::rgba(0.95f, 0.6f, 0.2f, 1.0f);
-static const Color kFieldEditUnder   = Color::rgba(0.95f, 0.6f, 0.2f, 0.9f);
-static const Color kFieldEditBg      = Color::rgba(0.95f, 0.6f, 0.2f, 0.18f);
-
-// Phase 3c — eyedropper cursor swatch chrome
-static const Color kEyedropChromeBg  = Color::rgba(0.08f, 0.08f, 0.1f, 0.95f);
-static const Color kEyedropBorder    = Color::rgba(1.0f, 1.0f, 1.0f, 0.85f);
-static const Color kEyedropText      = Color::rgba(0.95f, 0.95f, 1.0f, 1.0f);
-
-// Phase 3e — zoom loupe visual treatment
-static const Color kZoomPanelBg      = Color::rgba(0.06f, 0.06f, 0.08f, 0.97f);
-static const Color kZoomBorder       = Color::rgba(0.25f, 0.5f, 1.0f, 0.9f);
-static const Color kZoomGridLine     = Color::rgba(0.0f, 0.0f, 0.0f, 0.25f);
-static const Color kZoomCrosshair    = Color::rgba(1.0f, 0.5f, 0.0f, 0.95f);
-static const Color kZoomReadoutText  = Color::rgba(0.9f, 0.9f, 0.95f, 1.0f);
-static const Color kZoomReadoutDim   = Color::rgba(0.55f, 0.55f, 0.6f, 1.0f);
-// Two-tone checkerboard for the no-readback fallback grid.
-static const Color kZoomCheckerA     = Color::rgba(0.22f, 0.22f, 0.26f, 1.0f);
-static const Color kZoomCheckerB     = Color::rgba(0.30f, 0.30f, 0.34f, 1.0f);
 
 // Format a Color as a CSS hex string: "#rrggbb" when fully opaque,
 // "#rrggbbaa" otherwise. Lower-case, fixed-width — matches the hex
@@ -1532,218 +1502,9 @@ void InspectorOverlay::paint_stats_bar(Canvas& canvas, float x, float y, float w
     canvas.fill_text(std::to_string(count) + " views", x + w - 60, y + 16);
 }
 
-// ── Phase 6.1 — Per-pass GPU/render attribution viewer ──────────────────────
-//
-// Surfaces where render time goes, broken down by render pass, over a
-// rolling 60-frame window. Reads RenderPassManager's existing per-pass
-// PassStats — CPU wall-time + draw-call counts. True GPU timestamps are
-// deferred to Phase 6.5 (Dawn timestamp queries); the panel labels its
-// numbers "cpu" so the distinction is honest and explicit.
-
-namespace {
-
-// Stable human-readable names for the five RenderPassType values, in
-// declaration order so index == static_cast<size_t>(type).
-constexpr std::array<const char*, 5> kPassTypeNames = {
-    "background", "content", "effects", "overlay", "post"
-};
-
-// Color-code each pass type so the panel reads at a glance — cool hues
-// for the cheap structural passes, warm for the expensive ones.
-const std::array<Color, 5> kPassTypeColors = {
-    Color::rgba(0.40f, 0.55f, 0.85f, 1.0f),  // background — blue
-    Color::rgba(0.45f, 0.80f, 0.55f, 1.0f),  // content    — green
-    Color::rgba(0.90f, 0.70f, 0.30f, 1.0f),  // effects    — amber
-    Color::rgba(0.85f, 0.45f, 0.80f, 1.0f),  // overlay    — magenta
-    Color::rgba(0.90f, 0.45f, 0.35f, 1.0f),  // post       — red
-};
-
-} // namespace
-
-bool InspectorOverlay::capture_pass_frame() {
-    if (!rpm_) return false;
-
-    // De-dup: only capture once per render-pass-manager frame. paint()
-    // can run multiple times for the same frame (e.g. partial redraw),
-    // and we don't want those to fill the history with copies of one
-    // frame's numbers. A frame_count() of 0 means begin_frame() was
-    // never called — treat that as "no frame to capture yet".
-    const std::uint64_t frame = rpm_->frame_count();
-    if (frame == 0 || frame == last_captured_frame_) return false;
-    last_captured_frame_ = frame;
-    ++pass_frames_captured_;
-
-    if (rpm_->over_budget()) ++budget_overrun_count_;
-
-    // Mark every pass type absent for this frame; the loop below flips
-    // back on the ones that actually rendered.
-    for (auto& ring : pass_rings_) ring.present_last_frame = false;
-
-    // The manager can emit the same pass type more than once per frame
-    // (e.g. two overlay passes). Accumulate per type so the history
-    // sample is the frame's TOTAL cost for that pass type.
-    std::array<float, kPassTypeCount> frame_cpu_ms{};
-    std::array<int, kPassTypeCount> frame_draw_calls{};
-    std::array<bool, kPassTypeCount> frame_seen{};
-    for (const auto& p : rpm_->passes()) {
-        auto idx = static_cast<std::size_t>(p.type);
-        if (idx >= kPassTypeCount) continue;  // defensive — unknown enum.
-        frame_cpu_ms[idx] += p.time_ms;
-        frame_draw_calls[idx] += p.draw_calls;
-        frame_seen[idx] = true;
-    }
-
-    for (std::size_t i = 0; i < kPassTypeCount; ++i) {
-        if (!frame_seen[i]) continue;  // pass absent this frame — no sample.
-        auto& ring = pass_rings_[i];
-        ring.cpu_ms[ring.head] = frame_cpu_ms[i];
-        ring.draw_calls[ring.head] = frame_draw_calls[i];
-        ring.head = (ring.head + 1) % kPassHistoryFrames;
-        if (ring.count < kPassHistoryFrames) ++ring.count;
-        ring.present_last_frame = true;
-    }
-    return true;
-}
-
-std::vector<InspectorOverlay::PassAttribution>
-InspectorOverlay::pass_attribution() const {
-    std::vector<PassAttribution> out;
-    out.reserve(kPassTypeCount);
-    for (std::size_t i = 0; i < kPassTypeCount; ++i) {
-        const auto& ring = pass_rings_[i];
-        PassAttribution a;
-        a.type = static_cast<int>(i);
-        a.name = kPassTypeNames[i];
-        a.samples = ring.count;
-        a.present = ring.present_last_frame;
-        if (ring.count > 0) {
-            // The most recent sample sits one slot behind head (mod N).
-            std::size_t last = (ring.head + kPassHistoryFrames - 1) % kPassHistoryFrames;
-            a.last_cpu_ms = ring.cpu_ms[last];
-            a.last_draw_calls = ring.draw_calls[last];
-            float sum = 0.0f;
-            for (std::size_t k = 0; k < ring.count; ++k) {
-                float v = ring.cpu_ms[k];
-                sum += v;
-                if (v > a.peak_cpu_ms) a.peak_cpu_ms = v;
-                if (ring.draw_calls[k] > a.peak_draw_calls)
-                    a.peak_draw_calls = ring.draw_calls[k];
-            }
-            a.avg_cpu_ms = sum / static_cast<float>(ring.count);
-        }
-        out.push_back(a);
-    }
-    return out;
-}
-
-void InspectorOverlay::paint_pass_attribution(Canvas& canvas, float x, float y,
-                                              float w, float h) {
-    canvas.set_font("monospace", kFontSize);
-    float line_y = y + 4;
-    const float line_h = 15.0f;
-
-    // Heading + honesty note about CPU-vs-GPU timing.
-    canvas.set_fill_color(kHighlightStroke);
-    canvas.fill_text("Render Passes (P)", x, line_y + 11);
-    line_y += line_h;
-    canvas.set_fill_color(kPanelDim);
-    canvas.fill_text("cpu time \xc2\xb7 GPU timestamps: Phase 6.5", x, line_y + 10);
-    line_y += line_h + 2;
-
-    if (!rpm_) {
-        canvas.set_fill_color(kPanelDim);
-        canvas.fill_text("No RenderPassManager attached", x, line_y + 11);
-        return;
-    }
-
-    // Frame summary line: total CPU time, budget, overrun count.
-    {
-        std::ostringstream ss;
-        ss << std::fixed << std::setprecision(2)
-           << rpm_->total_time_ms() << "ms / "
-           << std::setprecision(1) << rpm_->budget() << "ms budget";
-        canvas.set_fill_color(rpm_->over_budget() ? kStatsWarn : kPanelText);
-        canvas.fill_text(ss.str(), x, line_y + 11);
-        line_y += line_h;
-
-        std::ostringstream ss2;
-        ss2 << budget_overrun_count_ << " overruns \xc2\xb7 "
-            << pass_frames_captured_ << " frames";
-        canvas.set_fill_color(kPanelDim);
-        canvas.fill_text(ss2.str(), x, line_y + 11);
-        line_y += line_h + 4;
-    }
-
-    auto attribution = pass_attribution();
-
-    // The trend sparkline scales relative to the worst single-pass
-    // CPU sample across all passes so bars are comparable to each other.
-    float global_peak = 0.001f;
-    for (const auto& a : attribution)
-        if (a.peak_cpu_ms > global_peak) global_peak = a.peak_cpu_ms;
-
-    bool any = false;
-    for (std::size_t i = 0; i < attribution.size(); ++i) {
-        const auto& a = attribution[i];
-        if (a.samples == 0) continue;  // pass never rendered — skip row.
-        any = true;
-        if (line_y > y + h - line_h * 2) break;  // out of panel space.
-
-        const Color& pass_color = kPassTypeColors[i];
-
-        // Color-coded pass-type chip + name. Dim the name when the pass
-        // was absent from the most recent frame (history but quiet now).
-        canvas.set_fill_color(pass_color);
-        canvas.fill_rounded_rect(x, line_y + 2, 8, 10, 2);
-        canvas.set_fill_color(a.present ? kPanelText : kPanelDim);
-        canvas.fill_text(a.name, x + 14, line_y + 11);
-
-        // last / avg / peak CPU ms, right-aligned-ish in a fixed column.
-        std::ostringstream stat;
-        stat << std::fixed << std::setprecision(2)
-             << a.last_cpu_ms << " ~" << a.avg_cpu_ms
-             << " ^" << a.peak_cpu_ms << "ms";
-        canvas.set_fill_color(kPanelDim);
-        canvas.fill_text(stat.str(), x + 90, line_y + 11);
-        line_y += line_h;
-
-        // Draw-call line.
-        std::ostringstream dc;
-        dc << a.last_draw_calls << " draws (peak " << a.peak_draw_calls << ")";
-        canvas.set_fill_color(kPanelDim);
-        canvas.fill_text(dc.str(), x + 14, line_y + 10);
-        line_y += line_h;
-
-        // 60-frame CPU-time sparkline. Walk the ring oldest→newest so
-        // the bars read left-to-right in chronological order.
-        const auto& ring = pass_rings_[i];
-        if (ring.count > 0) {
-            const float spark_h = 14.0f;
-            const float spark_w = std::min(w - 14.0f, 220.0f);
-            const float bar_w = spark_w / static_cast<float>(kPassHistoryFrames);
-            const float base_y = line_y + spark_h;
-            canvas.set_fill_color(Color::rgba(0, 0, 0, 0.35f));
-            canvas.fill_rect(x + 14, line_y, spark_w, spark_h);
-            std::size_t start = (ring.head + kPassHistoryFrames - ring.count)
-                                % kPassHistoryFrames;
-            for (std::size_t k = 0; k < ring.count; ++k) {
-                float v = ring.cpu_ms[(start + k) % kPassHistoryFrames];
-                float frac = v / global_peak;
-                if (frac > 1.0f) frac = 1.0f;
-                float bh = frac * spark_h;
-                canvas.set_fill_color(pass_color);
-                canvas.fill_rect(x + 14 + static_cast<float>(k) * bar_w,
-                                 base_y - bh, std::max(bar_w - 1.0f, 1.0f), bh);
-            }
-            line_y += spark_h + 6;
-        }
-    }
-
-    if (!any) {
-        canvas.set_fill_color(kPanelDim);
-        canvas.fill_text("Awaiting render frames\xe2\x80\xa6", x, line_y + 11);
-    }
-}
+// Phase 6.1 — the per-pass GPU/render attribution viewer
+// (capture_pass_frame / pass_attribution / paint_pass_attribution) is
+// defined in inspector_overlay_pass_viewer.cpp.
 
 // ── Phase 2.5 — Tweak management panel (Photoshop-layers style) ─────────────
 //
@@ -2298,537 +2059,13 @@ float InspectorOverlay::paint_drift_drawer(Canvas& canvas, float x, float y,
     return total_h;
 }
 
-// ── Phase 3b — Live-editable box-model fields ───────────────────────────────
-//
-// Click on a numeric value in the property panel → enter edit mode.
-// Typed digits / decimal / sign extend the buffer; arrows nudge (±1
-// plain, ±10 Shift, ±100 Cmd matching Figma); Enter commits (tweak
-// + persisted); Esc cancels (revert); Tab commits + moves to the
-// next editable field. The tweak is emitted via the SAME
-// emit_tweak_for_selection() path Phase 3a drag-handles use, so
-// downstream persistence (Phase 1 disk write) gets both gestures for
-// free.
-//
-// Real-time preview: we write to the underlying View (flex().padding
-// = N etc.) on every keystroke + arrow nudge so Yoga reflows live —
-// matching the spec's "updates in real-time as Yoga reflow runs"
-// requirement. The tweak (anchor-keyed persisted edit) is emitted
-// only on commit, so a user who Esc-cancels never poisons the store.
+// Phase 3b — the live-editable box-model fields (editable_field_at /
+// read_field_value / write_field_value / begin_field_edit /
+// commit_field_edit / cancel_field_edit / apply_edit_buffer_to_view /
+// handle_edit_key) are defined in inspector_overlay_field_edit.cpp.
 
-int InspectorOverlay::editable_field_at(Point p) const {
-    for (std::size_t i = 0; i < editable_fields_.size(); ++i) {
-        const auto& f = editable_fields_[i];
-        if (p.x >= f.bounds.x && p.x <= f.bounds.x + f.bounds.width &&
-            p.y >= f.bounds.y && p.y <= f.bounds.y + f.bounds.height)
-            return static_cast<int>(i);
-    }
-    return -1;
-}
-
-float InspectorOverlay::read_field_value(std::string_view field_path) const {
-    if (!selected_) return 0.0f;
-    const auto& f = selected_->flex();
-    if (field_path == "layout.width")    return f.preferred_width;
-    if (field_path == "layout.height")   return f.preferred_height;
-    if (field_path == "layout.padding")  return f.padding;
-    if (field_path == "layout.margin")   return f.margin;
-    if (field_path == "style.opacity")   return selected_->opacity();
-    // style.font_size is documented in the spec but View has no
-    // direct font_size accessor at the View level — Label / TextEditor
-    // own their own font sizes. Skip until widget-aware property
-    // mapping ships (noted in PR body).
-    return 0.0f;
-}
-
-void InspectorOverlay::write_field_value(std::string_view field_path, float value) {
-    if (!selected_) return;
-    auto& f = selected_->flex();
-    bool layout_touched = false;
-    if (field_path == "layout.width") {
-        f.preferred_width = value; layout_touched = true;
-    } else if (field_path == "layout.height") {
-        f.preferred_height = value; layout_touched = true;
-    } else if (field_path == "layout.padding") {
-        f.padding = value; layout_touched = true;
-    } else if (field_path == "layout.margin") {
-        f.margin = value; layout_touched = true;
-    } else if (field_path == "style.opacity") {
-        // Opacity clamps to [0,1] inside View::set_opacity.
-        selected_->set_opacity(value);
-    }
-    if (layout_touched) {
-        selected_->invalidate_layout();
-        // Yoga propagates from the dirty node up to the next absolute-
-        // position container or the root; mark up to the root so the
-        // next paint pass definitely recomputes. (Cheap — the flag is
-        // just a bool.)
-        for (View* v = selected_; v; v = v->parent())
-            v->invalidate_layout();
-    }
-}
-
-bool InspectorOverlay::begin_field_edit(std::string field_path, float initial_value) {
-    if (!selected_) return false;
-    if (field_path.empty()) return false;
-    editing_field_ = std::move(field_path);
-    edit_target_view_ = selected_;
-    edit_original_value_ = initial_value;
-    // Format as integer when the value is whole-number-ish; this
-    // matches the static display logic in paint_props_section so the
-    // user sees the same digits they were looking at a frame ago.
-    std::ostringstream oss;
-    if (std::abs(initial_value - std::round(initial_value)) < 1e-4f) {
-        oss << static_cast<int>(std::round(initial_value));
-    } else {
-        oss << std::fixed << std::setprecision(2) << initial_value;
-    }
-    edit_buffer_ = oss.str();
-    edit_caret_pos_ = edit_buffer_.size();
-    return true;
-}
-
-bool InspectorOverlay::commit_field_edit() {
-    if (editing_field_.empty()) return false;
-
-    // Parse buffer; tolerate trailing whitespace / empty (treat empty
-    // as "no change" — revert without emitting a tweak).
-    bool emitted = false;
-    if (!edit_buffer_.empty()) {
-        char* end = nullptr;
-        double parsed = std::strtod(edit_buffer_.c_str(), &end);
-        if (end != edit_buffer_.c_str()) {
-            float value = static_cast<float>(parsed);
-            // Real-time preview already applied on each keystroke, but
-            // ensure the final committed value is correct (in case the
-            // last keystroke wasn't a digit).
-            write_field_value(editing_field_, value);
-            // Emit the tweak — persisted edit, anchor-keyed.
-            emit_tweak_for_selection(editing_field_,
-                                     choc::value::createFloat32(value),
-                                     "inspector-keyboard-edit");
-            emitted = true;
-        }
-    }
-
-    editing_field_.clear();
-    edit_buffer_.clear();
-    edit_caret_pos_ = 0;
-    edit_target_view_ = nullptr;
-    return emitted;
-}
-
-void InspectorOverlay::cancel_field_edit() {
-    if (editing_field_.empty()) return;
-    // Revert the underlying View to the original value — real-time
-    // preview may have mutated it.
-    write_field_value(editing_field_, edit_original_value_);
-    editing_field_.clear();
-    edit_buffer_.clear();
-    edit_caret_pos_ = 0;
-    edit_target_view_ = nullptr;
-}
-
-void InspectorOverlay::apply_edit_buffer_to_view() {
-    if (editing_field_.empty()) return;
-    if (edit_buffer_.empty()) return;
-    char* end = nullptr;
-    double parsed = std::strtod(edit_buffer_.c_str(), &end);
-    if (end == edit_buffer_.c_str()) return;  // not a number yet
-    write_field_value(editing_field_, static_cast<float>(parsed));
-}
-
-// Translate a KeyCode into the digit / sign / decimal character it
-// would produce in a US-layout context. Shift is intentionally
-// ignored — the spec only needs unmodified digits + period + minus.
-// (Full keyboard mapping is platform-layer concern; this is a
-// pragmatic subset for box-model numeric entry.)
-static char key_to_char(KeyCode k, bool shift) {
-    int v = static_cast<int>(k);
-    if (v >= '0' && v <= '9') return static_cast<char>(v);
-    // We treat the keys as both their unshifted and shifted symbols
-    // for "." and "-" because Pulp's KeyCode enum doesn't define
-    // separate "period" / "minus" entries — platform code dispatches
-    // them via text-input. Phase 3b accepts numeric editing only;
-    // digits cover 99% of use. Decimal entry on platforms without a
-    // text-input path can be added by extending KeyCode (filed as a
-    // follow-up in the PR body if it becomes a real ask).
-    (void)shift;
-    return 0;
-}
-
-bool InspectorOverlay::handle_edit_key(const KeyEvent& event) {
-    // ── Cancel: Esc ────────────────────────────────────────────────
-    if (event.key == KeyCode::escape) {
-        cancel_field_edit();
-        return true;
-    }
-
-    // ── Commit: Enter ──────────────────────────────────────────────
-    if (event.key == KeyCode::enter) {
-        commit_field_edit();
-        return true;
-    }
-
-    // ── Tab: commit + move to next field ──────────────────────────
-    if (event.key == KeyCode::tab) {
-        // Find current field index, then move +1 (Shift+Tab = -1)
-        // among the editable_fields_ list (populated by the last
-        // paint). The list is repopulated each paint so it's a stable
-        // ordering matching what the user sees.
-        std::string current = editing_field_;
-        commit_field_edit();
-
-        int idx = -1;
-        for (std::size_t i = 0; i < editable_fields_.size(); ++i) {
-            if (editable_fields_[i].path == current) { idx = static_cast<int>(i); break; }
-        }
-        if (idx >= 0 && !editable_fields_.empty()) {
-            int step = event.isShiftDown() ? -1 : 1;
-            int next = (idx + step) % static_cast<int>(editable_fields_.size());
-            if (next < 0) next += static_cast<int>(editable_fields_.size());
-            const auto& nf = editable_fields_[next];
-            // Re-read the value: previous commit may have changed it.
-            float v = read_field_value(nf.path);
-            begin_field_edit(nf.path, v);
-        }
-        return true;
-    }
-
-    // ── Backspace: trim one char ──────────────────────────────────
-    if (event.key == KeyCode::backspace) {
-        if (edit_caret_pos_ > 0 && !edit_buffer_.empty()) {
-            edit_buffer_.erase(edit_caret_pos_ - 1, 1);
-            --edit_caret_pos_;
-            apply_edit_buffer_to_view();
-        }
-        return true;
-    }
-
-    // ── Arrow nudging: ±1 / ±10 (shift) / ±100 (cmd) ──────────────
-    if (event.key == KeyCode::up || event.key == KeyCode::down) {
-        float step = 1.0f;
-        if (event.isShiftDown()) step = 10.0f;
-        if (event.isMainModifier()) step = 100.0f;
-        if (event.key == KeyCode::down) step = -step;
-
-        // Parse current buffer + apply step. Fall back to original
-        // value if parse fails (e.g. buffer is empty / only "-").
-        float current = edit_original_value_;
-        if (!edit_buffer_.empty()) {
-            char* end = nullptr;
-            double parsed = std::strtod(edit_buffer_.c_str(), &end);
-            if (end != edit_buffer_.c_str())
-                current = static_cast<float>(parsed);
-        }
-        float new_v = current + step;
-
-        // Re-format the buffer; keep integer formatting if both ends
-        // are whole numbers so we don't surprise the user with
-        // suddenly-decimal values from arrow nudging.
-        std::ostringstream oss;
-        if (std::abs(new_v - std::round(new_v)) < 1e-4f &&
-            std::abs(current - std::round(current)) < 1e-4f) {
-            oss << static_cast<int>(std::round(new_v));
-        } else {
-            oss << std::fixed << std::setprecision(2) << new_v;
-        }
-        edit_buffer_ = oss.str();
-        edit_caret_pos_ = edit_buffer_.size();
-        apply_edit_buffer_to_view();
-        return true;
-    }
-
-    // ── Left / Right: caret movement ──────────────────────────────
-    if (event.key == KeyCode::left) {
-        if (edit_caret_pos_ > 0) --edit_caret_pos_;
-        return true;
-    }
-    if (event.key == KeyCode::right) {
-        if (edit_caret_pos_ < edit_buffer_.size()) ++edit_caret_pos_;
-        return true;
-    }
-
-    // ── Home / End ────────────────────────────────────────────────
-    if (event.key == KeyCode::home) {
-        edit_caret_pos_ = 0;
-        return true;
-    }
-    if (event.key == KeyCode::end_) {
-        edit_caret_pos_ = edit_buffer_.size();
-        return true;
-    }
-
-    // ── Digit insertion ───────────────────────────────────────────
-    // Refuse digits when a modifier (Cmd / Ctrl) is held so we don't
-    // swallow chord shortcuts.
-    if (event.isCmdDown() || event.isCtrlDown()) return false;
-
-    char c = key_to_char(event.key, event.isShiftDown());
-    if (c >= '0' && c <= '9') {
-        edit_buffer_.insert(edit_caret_pos_, 1, c);
-        ++edit_caret_pos_;
-        apply_edit_buffer_to_view();
-        return true;
-    }
-
-    return false;
-}
-
-// ── Phase 3e — 20× zoom loupe ───────────────────────────────────────────────
-//
-// A digital loupe: a fixed-corner panel showing the pixels under the
-// cursor blown up by `zoom_factor_`, with a center crosshair on the
-// exact sample pixel and a coordinate + hex readout. It pairs with the
-// eyedropper — the eyedropper grabs one pixel, the loupe shows the
-// neighborhood so edge alignment and color boundaries are visible.
-
-void InspectorOverlay::set_zoom_active(bool active) {
-    zoom_active_ = active;
-    if (active) {
-        // Seed the sample center so the first paint has a sane region
-        // even before the cursor moves — the panel center is a safe,
-        // always-on-screen default.
-        zoom_sample_center_ = {root_.bounds().width * 0.5f,
-                               root_.bounds().height * 0.5f};
-    }
-}
-
-void InspectorOverlay::set_zoom_factor(int factor) {
-    zoom_factor_ = std::clamp(factor, kZoomFactorMin, kZoomFactorMax);
-}
-
-// Walk the view tree top-most-first; return the deepest View that both
-// contains `p` AND carries an explicit background color. This is the
-// no-readback fallback: when read_pixels() isn't available we paint the
-// loupe with the resolved authored colors instead of true device pixels.
-bool InspectorOverlay::resolve_view_color_at(Point p, Color& out) const {
-    const View* best = nullptr;
-    std::function<void(const View*)> walk = [&](const View* v) {
-        if (!v) return;
-        auto r = view_bounds_in_root(v);
-        bool inside = p.x >= r.x && p.x < r.x + r.width &&
-                      p.y >= r.y && p.y < r.y + r.height;
-        if (inside && v->has_background_color()) best = v;
-        // Children paint over parents — visiting them after the parent
-        // means a deeper hit overwrites `best`, matching paint order.
-        for (size_t i = 0; i < v->child_count(); ++i) walk(v->child_at(i));
-    };
-    walk(&root_);
-    if (!best) return false;
-    out = best->background_color();
-    return true;
-}
-
-// Refresh zoom_center_color_ for the current sample center. Tries a
-// real pixel readback first (Skia raster only); falls back to resolved
-// view color, and records which path ran in zoom_center_from_readback_
-// so the readout can label degraded samples honestly.
-void InspectorOverlay::update_zoom_sample(Canvas& canvas) {
-    // The overlay paints onto a surface sized to the root view, so the
-    // root bounds are the readback clamp limits. Clamp the sample pixel
-    // into [0, w-1] × [0, h-1] — a raw read at an off-canvas coord (the
-    // cursor sitting on the very edge) fails outright, which would push
-    // the readout onto the degraded fallback path even on a readback-
-    // capable surface. Clamping keeps the center readout honest about
-    // readback at edges, and keeps it consistent with the block read in
-    // paint_zoom_panel() (which clamps to the same bounds).
-    const int cw = static_cast<int>(root_.bounds().width);
-    const int ch = static_cast<int>(root_.bounds().height);
-    int cx = static_cast<int>(zoom_sample_center_.x);
-    int cy = static_cast<int>(zoom_sample_center_.y);
-    if (cw > 0) cx = std::clamp(cx, 0, cw - 1);
-    if (ch > 0) cy = std::clamp(cy, 0, ch - 1);
-
-    std::uint8_t rgba[4] = {0, 0, 0, 0};
-    if (canvas.read_pixels(cx, cy, 1, 1, rgba)) {
-        zoom_center_color_ = Color::rgba(rgba[0] / 255.0f, rgba[1] / 255.0f,
-                                         rgba[2] / 255.0f, rgba[3] / 255.0f);
-        zoom_center_from_readback_ = true;
-        return;
-    }
-    // No pixel readback on this surface (RecordingCanvas, CG fallback,
-    // headless). Degrade gracefully to the authored view color.
-    Color resolved{};
-    if (resolve_view_color_at(zoom_sample_center_, resolved)) {
-        zoom_center_color_ = resolved;
-    } else {
-        zoom_center_color_ = Color::rgba(0, 0, 0, 0);  // nothing under cursor
-    }
-    zoom_center_from_readback_ = false;
-}
-
-void InspectorOverlay::paint_zoom_panel(Canvas& canvas) {
-    // Resolve the center pixel + degradation state for this frame.
-    update_zoom_sample(canvas);
-
-    const int   cells = kZoomGridCells;          // odd → exact center cell
-    const int   half  = cells / 2;
-    const float cell  = static_cast<float>(zoom_factor_);
-    const float grid_px = cell * cells;
-
-    // ── Panel placement: fixed bottom-left corner. Fixed (vs cursor-
-    // following) avoids flicker and never occludes the cursor target;
-    // bottom-left stays clear of the props panel on the right.
-    const float pad = 6.0f;
-    const float panel_w = grid_px + pad * 2.0f;
-    const float panel_h = grid_px + pad * 2.0f + kZoomReadoutH;
-    const float panel_x = kZoomPanelMargin;
-    const float panel_y = root_.bounds().height - panel_h - kZoomPanelMargin;
-    const float grid_x  = panel_x + pad;
-    const float grid_y  = panel_y + pad;
-
-    canvas.save();
-
-    // Panel background + border.
-    canvas.set_fill_color(kZoomPanelBg);
-    canvas.fill_rect(panel_x, panel_y, panel_w, panel_h);
-    canvas.set_stroke_color(kZoomBorder);
-    canvas.set_line_width(1.5f);
-    canvas.stroke_rect(panel_x, panel_y, panel_w, panel_h);
-
-    // ── Magnified pixel grid ────────────────────────────────────────
-    // Each cell maps to one source pixel. With a real readback we ask
-    // the canvas for the whole NxN block at once; otherwise we
-    // synthesize a checkerboard (with the resolved center color in the
-    // middle) so the loupe still communicates "this is where you're
-    // sampling".
-    //
-    // Edge clamping (codex P2 #2464): a window centered on a cursor
-    // within `half` pixels of any canvas edge would put `ox`/`oy`
-    // negative or push `ox+cells`/`oy+cells` past the surface — Skia's
-    // readPixels() rejects an out-of-bounds source rect outright, so
-    // the WHOLE block dropped to checkerboard exactly where pixel
-    // inspection matters most. Clamp the read origin so the full
-    // `cells × cells` window stays in-bounds; the magnified region
-    // shifts slightly near edges but still shows real device pixels.
-    // The sample pixel may then sit off-center within the grid, so the
-    // crosshair below tracks its actual cell (cross_gx/cross_gy)
-    // instead of assuming the middle.
-    const int   cw = static_cast<int>(root_.bounds().width);
-    const int   ch = static_cast<int>(root_.bounds().height);
-    int ox = static_cast<int>(zoom_sample_center_.x) - half;
-    int oy = static_cast<int>(zoom_sample_center_.y) - half;
-    if (cw >= cells) ox = std::clamp(ox, 0, cw - cells);
-    if (ch >= cells) oy = std::clamp(oy, 0, ch - cells);
-
-    // The sample pixel's cell within the (possibly clamped) grid. When
-    // the window isn't clamped this is the exact center (half, half);
-    // near an edge it shifts toward the edge the cursor approached.
-    int sample_px = std::clamp(static_cast<int>(zoom_sample_center_.x),
-                               0, cw > 0 ? cw - 1 : 0);
-    int sample_py = std::clamp(static_cast<int>(zoom_sample_center_.y),
-                               0, ch > 0 ? ch - 1 : 0);
-    const int cross_gx = std::clamp(sample_px - ox, 0, cells - 1);
-    const int cross_gy = std::clamp(sample_py - oy, 0, cells - 1);
-
-    std::vector<std::uint8_t> block(static_cast<size_t>(cells) * cells * 4, 0);
-    const bool have_block = canvas.read_pixels(ox, oy, cells, cells,
-                                               block.data());
-
-    for (int gy = 0; gy < cells; ++gy) {
-        for (int gx = 0; gx < cells; ++gx) {
-            Color c;
-            if (have_block) {
-                const size_t i = (static_cast<size_t>(gy) * cells + gx) * 4;
-                c = Color::rgba(block[i] / 255.0f, block[i + 1] / 255.0f,
-                                block[i + 2] / 255.0f, block[i + 3] / 255.0f);
-            } else if (gx == cross_gx && gy == cross_gy) {
-                // Center (sample) cell shows the resolved sample color
-                // even on the fallback path so the readout and grid
-                // agree — and so they agree about WHICH cell is the
-                // sample pixel when the window is edge-clamped.
-                c = zoom_center_color_;
-            } else {
-                c = ((gx + gy) & 1) ? kZoomCheckerB : kZoomCheckerA;
-            }
-            canvas.set_fill_color(c);
-            canvas.fill_rect(grid_x + gx * cell, grid_y + gy * cell,
-                             cell, cell);
-        }
-    }
-
-    // ── Pixel grid lines ────────────────────────────────────────────
-    // Thin lines between magnified pixels so individual pixels read as
-    // discrete cells, like Photoshop's pixel-grid at high zoom.
-    canvas.set_stroke_color(kZoomGridLine);
-    canvas.set_line_width(1.0f);
-    for (int i = 0; i <= cells; ++i) {
-        const float gx = grid_x + i * cell;
-        const float gyl = grid_y + i * cell;
-        canvas.stroke_line(gx, grid_y, gx, grid_y + grid_px);
-        canvas.stroke_line(grid_x, gyl, grid_x + grid_px, gyl);
-    }
-
-    // ── Center crosshair / target ───────────────────────────────────
-    // Outlines the exact sample pixel — the cell the readout describes.
-    // Uses cross_gx/cross_gy (not `half`) so the crosshair still marks
-    // the true sample pixel when the read window was edge-clamped.
-    const float center_x = grid_x + cross_gx * cell;
-    const float center_y = grid_y + cross_gy * cell;
-    canvas.set_stroke_color(kZoomCrosshair);
-    canvas.set_line_width(2.0f);
-    canvas.stroke_rect(center_x, center_y, cell, cell);
-    // Crosshair arms reaching from the grid edges toward the center
-    // cell, so the eye is guided to the sample pixel.
-    const float cxm = center_x + cell * 0.5f;
-    const float cym = center_y + cell * 0.5f;
-    canvas.set_line_width(1.0f);
-    canvas.stroke_line(grid_x, cym, center_x, cym);
-    canvas.stroke_line(center_x + cell, cym, grid_x + grid_px, cym);
-    canvas.stroke_line(cxm, grid_y, cxm, center_y);
-    canvas.stroke_line(cxm, center_y + cell, cxm, grid_y + grid_px);
-
-    // ── Coordinate + hex readout strip ──────────────────────────────
-    const float readout_y = grid_y + grid_px + 2.0f;
-    canvas.set_font("monospace", 10.0f);
-
-    const int sx = static_cast<int>(zoom_sample_center_.x);
-    const int sy = static_cast<int>(zoom_sample_center_.y);
-    std::ostringstream coord;
-    coord << zoom_factor_ << "x  (" << sx << ", " << sy << ")";
-    canvas.set_fill_color(kZoomReadoutText);
-    canvas.fill_text(coord.str(), grid_x, readout_y + 12.0f);
-
-    // Hex color readout — #RRGGBB plus an alpha byte when not opaque.
-    auto to_byte = [](float v) {
-        return static_cast<int>(std::lround(std::clamp(v, 0.0f, 1.0f) * 255.0f));
-    };
-    const int hr = to_byte(zoom_center_color_.r);
-    const int hg = to_byte(zoom_center_color_.g);
-    const int hb = to_byte(zoom_center_color_.b);
-    const int ha = to_byte(zoom_center_color_.a);
-    std::ostringstream hex;
-    hex << '#' << std::hex << std::uppercase << std::setfill('0')
-        << std::setw(2) << hr << std::setw(2) << hg << std::setw(2) << hb;
-    if (ha != 255) hex << std::setw(2) << ha;
-
-    // Color swatch + hex text on the second readout line.
-    const float swatch = 10.0f;
-    const float swatch_y = readout_y + 16.0f;
-    canvas.set_fill_color(zoom_center_color_);
-    canvas.fill_rect(grid_x, swatch_y, swatch, swatch);
-    canvas.set_stroke_color(kZoomReadoutDim);
-    canvas.set_line_width(1.0f);
-    canvas.stroke_rect(grid_x, swatch_y, swatch, swatch);
-    canvas.set_fill_color(kZoomReadoutText);
-    canvas.fill_text(hex.str(), grid_x + swatch + 6.0f, swatch_y + 9.0f);
-
-    // Honesty label: tell the user when the magnified grid is a
-    // fallback render rather than true device pixels, so they don't
-    // trust a checker pattern as real output. Keyed off `have_block`
-    // (the grid's actual readback result) rather than just the center
-    // pixel's `zoom_center_from_readback_` — with edge clamping the
-    // two now agree on any normal surface, but on a degenerate canvas
-    // narrower than the grid the 1×1 center read can still succeed
-    // while the cells×cells block cannot, and the label must describe
-    // the grid the user is looking at.
-    if (!have_block) {
-        canvas.set_fill_color(kZoomReadoutDim);
-        canvas.fill_text("no readback", grid_x + grid_px - 64.0f,
-                         swatch_y + 9.0f);
-    }
-
-    canvas.restore();
-}
+// Phase 3e — the 20× zoom loupe (set_zoom_active / set_zoom_factor /
+// resolve_view_color_at / update_zoom_sample / paint_zoom_panel) is
+// defined in inspector_overlay_zoom.cpp.
 
 } // namespace pulp::inspect
